TR_OCR/requirements.txt
matplotlib==3.6.2
numpy==1.23.4
opencv_contrib_python_headless==4.6.0.66
Pillow==9.3.0
transformers==4.24.0
